+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to `laravel-cognito-auth` will be documented in this file
 
+## 0.1.3 - 2018-08-20
+
+- implemented confirmSignUp function in CognitoClient
+- added CognitoUserPropertyAccessor class 
+- added delete_user config
 
 ## 0.1.2 - 2018-08-19